requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
aiohttp==3.8.5
pillow==10.0.0
requests==2.31.0
praw==7.7.1
asyncpraw==7.7.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1 